change datatype from Double to BigDecimal

+++ src/main/java/com/business/keeper/exchangerateservice/entity/HistoryEntity.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Entity
@@ -25,6 +26,6 @@ public class HistoryEntity {
     private Currency base;
     private Currency target;
     private Double exchangeRate;
-    private Double average;
+    private BigDecimal average;
     private RateTrend trend;
 }
