+++ src/main/java/com/business/keeper/exchangerateservice/model/ExchangeRateResultDto.java
@@ -3,6 +3,8 @@ package com.business.keeper.exchangerateservice.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ExchangeRateResultDto {
@@ -10,7 +12,7 @@ public class ExchangeRateResultDto {
     private Currency base;
     private Currency target;
     private Double exchangeRate;
-    private Double average;
+    private BigDecimal average;
     private RateTrend trend;
     
 }
