+++ src/main/java/com/business/keeper/exchangerateservice/service/ExchangeRateService.java
@@ -2,6 +2,7 @@ package com.business.keeper.exchangerateservice.service;
 
 import com.business.keeper.exchangerateservice.model.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 import java.util.TreeMap;
@@ -9,7 +10,7 @@ import java.util.TreeMap;
 public interface ExchangeRateService {
     ExchangeRateResultDto getExchangeRateForGivenDateAndCurrencies(String date, Currency baseCurrency, Currency targetCurrency);
 
-    Double calculateAverage(HistoryExchangeRateAPIResponse historyExchangeRateAPIResponse, Currency target);
+    BigDecimal calculateAverage(HistoryExchangeRateAPIResponse historyExchangeRateAPIResponse, Currency target);
 
-    RateTrend determineRateTrend(TreeMap<LocalDate, Map<Currency, Double>> rates, Currency target);
+    RateTrend determineRateTrend(TreeMap<LocalDate, Map<Currency, BigDecimal>> rates, Currency target);
 }
